backend/requirements.txt
Django >= 3.1
djangorestframework ~= 3.12
django-cors-headers ~= 3.5
google-cloud-storage ~=1.31.2